--- modules/util/bundle-support/bnd.bnd
@@ -1,20 +1,0 @@
-Bundle-Name: Liferay Bundle Support
-Bundle-SymbolicName: com.liferay.bundle.support
-Bundle-Version: 1.0.0
-Export-Package:\
-	com.liferay.bundle.support,\
-	com.liferay.bundle.support.ant,\
-	com.liferay.bundle.support.maven
-Import-Package:\
-	!org.apache.maven.plugin.*,\
-	!org.apache.tools.ant.*,\
-	*
-JPM-Command: liferay-bundle-support
-Main-Class: com.liferay.bundle.support.BundleSupport
--includeresource:\
-	@commons-codec-[0-9]*.jar,\
-	@commons-compress-[0-9]*.jar,\
-	@commons-logging-[0-9]*.jar,\
-	@httpclient-[0-9]*.jar,\
-	@httpcore-[0-9]*.jar,\
-	@jcommander-[0-9]*.jar
