--- modules/util/bundle-support/samples/build.xml
@@ -1,26 +1,0 @@
-<?xml version="1.0"?>
-<!DOCTYPE project>
-
-<project>
-	<path id="deployment.helper.classpath">
-		<fileset dir="../../../../tools/sdk/dist" includes="com.liferay.bundle.support-*.jar" />
-	</path>
-
-	<taskdef classpathref="bundle.support.classpath" resource="com/liferay/bundle/support/ant/taskdefs.properties" />
-
-	<target name="clean">
-		<clean fileName="${plugin.file}" liferayHomeDir="${app.server.parent.dir}" />
-	</target>
-
-	<target name="deploy">
-		<deploy file="${plugin.file}" liferayHomeDir="${app.server.parent.dir}" outputFileName="${plugin.file}" />
-	</target>
-
-	<target name="dist-bundle">
-		<dist-bundle format="zip" includeFolder="true" liferayHomeDir="${app.server.parent.dir}" outputFile="${plugins.sdk.name}" />
-	</target>
-
-	<target name="init-bundle">
-		<init-bundle liferayHomeDir="${app.server.parent.dir}" />
-	</target>
-</project>
