--- modules/util/bundle-support/src/main/resources/com/liferay/bundle/support/ant/taskdefs.properties
@@ -1,4 +1,0 @@
-clean=com.liferay.bundle.support.ant.CleanTask
-deploy=com.liferay.bundle.support.ant.DeployTask
-dist-bundle=com.liferay.bundle.support.ant.DistBundleTask
-init-bundle=com.liferay.bundle.support.ant.InitBundleTask
