+++ modules/util/portal-tools-bundle-support/samples/pom.xml
@@ -1,0 +1,53 @@
+<?xml version="1.0"?>
+
+<project
+	xmlns="http://maven.apache.org/POM/4.0.0"
+	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd"
+>
+	<modelVersion>4.0.0</modelVersion>
+	<groupId>sample.deployment.helper</groupId>
+	<artifactId>sample.deployment.helper</artifactId>
+	<name>Sample Deployment Helper Project</name>
+	<version>1.0.0</version>
+	<description>Sample Deployment Helper Project</description>
+	<build>
+		<plugins>
+			<plugin>
+				<groupId>com.liferay</groupId>
+				<artifactId>com.liferay.portal.tools.bundle.support</artifactId>
+				<version>1.0.0</version>
+				<executions>
+					<execution>
+						<id>default-clean</id>
+						<goals>
+							<goal>clean</goal>
+						</goals>
+						<phase>clean</phase>
+					</execution>
+					<execution>
+						<id>default-init-bundle</id>
+						<goals>
+							<goal>init-bundle</goal>
+						</goals>
+						<phase>initialize</phase>
+					</execution>
+					<execution>
+						<id>default-deploy</id>
+						<goals>
+							<goal>deploy</goal>
+						</goals>
+						<phase>pre-integration-test</phase>
+					</execution>
+					<execution>
+						<id>default-dist-bundle</id>
+						<goals>
+							<goal>dist-bundle</goal>
+						</goals>
+						<phase>post-integration-test</phase>
+					</execution>
+				</executions>
+			</plugin>
+		</plugins>
+	</build>
+</project>
