+++ modules/util/portal-tools-bundle-support/src/main/resources/META-INF/m2e/lifecycle-mapping-metadata.xml
