+++ modules/util/portal-tools-bundle-support/src/main/resources/com/liferay/portal/tools/bundle/support/ant/taskdefs.properties
@@ -1,0 +1,4 @@
+clean=com.liferay.portal.tools.bundle.support.ant.CleanTask
+deploy=com.liferay.portal.tools.bundle.support.ant.DeployTask
+dist-bundle=com.liferay.portal.tools.bundle.support.ant.DistBundleTask
+init-bundle=com.liferay.portal.tools.bundle.support.ant.InitBundleTask
