+++ modules/util/portal-tools-bundle-support/src/test/resources/com/liferay/portal/tools/bundle/support/dependencies/test.zip
